refactor: 2개 이상의 클래스가 있는 파일들의 클래스를 별도의 파일로 분리
