+++ app/src/main/java/woowacourse/omok/Room.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok
+
+data class Room(
+    val id: Long,
+    val title: String,
+    val status: String,
+)
