+++ app/src/main/java/woowacourse/omok/model/database/DatabaseDao.kt
@@ -1,0 +1,11 @@
+package woowacourse.omok.model.database
+
+import android.database.sqlite.SQLiteOpenHelper
+
+interface DatabaseDao<T> {
+    val dbHelper: SQLiteOpenHelper
+
+    fun save(item: T): T
+
+    fun drop()
+}
