+++ app/src/main/java/woowacourse/omok/model/database/Placement.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.model.database
+
+data class Placement(
+    val placementId: Long = 0L,
+    val gameId: Long,
+    val color: String?,
+    val index: Int,
+)
